RAG-Agent-For-Website/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
langchain>=0.3.25
langchain-community>=0.3.24
langchain-core>=0.3.59
pydantic>=2.11.4
sentence-transformers
faiss-cpu
python-dotenv
langchain-huggingface